- proj_func unit is leaved as is (not converted to Ha). I don't really   know the actual unit used. - Dion is now assumed to be in Ry, so the factor of 0.5 follows   instead of 2.
Previously unit of proj_func and Dion is deducted by
comparing analytical GTH parameters and numeric GTH UPF. This
lead to the previously used conversion factor, however I found
that in USPP case I need factor of 4 in Qfunc.
This factor of 4 is however not working in case of PAW as the
cancellation is not working (units of ddd_paw, becsum, PAW_rho_lm, etc).

- Implement radial cutoff when constructing V_Ps_loc

- Some clean up